Applied fixes
