+++ src/test/ua/biz/test/ApplicationTest.java
@@ -5,19 +5,36 @@ import org.testng.Assert;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
-import ua.biz.test.config.Main;
+import ua.biz.test.config.EmployeeSQLDataProvider;
+import ua.biz.test.config.VacancySQLDataProvider;
 import ua.biz.test.page.LoginPage;
+import ua.biz.test.pojo.Employee;
+import ua.biz.test.pojo.JobVacancy;
 import ua.biz.test.util.LocalDriver;
 
+import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 public class ApplicationTest {
     private static final String URL = "http://test.biz.ua";
     private String jobVacancyName = "Test";
-    public String name = "Bobbie";
-    public String lastName = "Abernathy";
+    private String name = "Bobbie";
+    private String lastName = "Abernathy";
+    private EmployeeSQLDataProvider employeeSQLDataProvider =
+            new EmployeeSQLDataProvider()
+                    .fetchEmployeeList();
 
-    Main main = new Main().setEmployeeList(jobVacancyName).setVacancyList(name, lastName);
+    private VacancySQLDataProvider vacancySQLDataProvider =
+            new VacancySQLDataProvider()
+                    .fetchVacancyList(employeeSQLDataProvider);
+
+    private List<Employee> employees =
+            employeeSQLDataProvider
+                    .getEmployeesByJobVacancyName(jobVacancyName);
+    
+    private List<JobVacancy> vacancies =
+            vacancySQLDataProvider
+                    .getVacanciesByEmployeeName(name, lastName, employeeSQLDataProvider);
 
     @BeforeMethod
     public void setUp() {
@@ -34,7 +51,7 @@ public class ApplicationTest {
                         .selectRecruitment()
                         .switchToVacancy()
                         .switchToVacancy()
-                        .isVacancyInList(name, lastName, main.getVacancies());
+                        .isVacancyInList(name, lastName, vacancies);
 
         Assert.assertTrue(isVacancy, "Vacancy is not found!");
     }
@@ -45,7 +62,7 @@ public class ApplicationTest {
                 new LoginPage()
                         .loginIntoSystem()
                         .selectPim()
-                        .isEmployeeInList(main.getEmployees());
+                        .isEmployeeInList(employees);
 
         Assert.assertTrue(isEmployee, "Employee is not found!");
     }
